pyproject.toml
[tool.poetry]
name = "addressbookapi"
version = "0.1.0"
description = ""
authors = ["GodVov4 <volume245@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
fastapi = "0.105.0"
uvicorn = {extras = ["standard"], version = "0.24.0.post1"}
sqlalchemy = "2.0.23"
alembic = "1.13.0"
asyncpg = "0.29.0"
pydantic = {extras = ["email"], version = "2.5.2"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-jose = "3.3.0"
libgravatar = "1.0.4"
python-multipart = "0.0.6"
fastapi-users = {extras = ["sqlalchemy"], version = "12.1.2"}
fastapi-mail = "1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
